+++ R/nesting.R
@@ -8,6 +8,16 @@
 #' `NULL` then it is added to all.
 #' @param .data A dataset to use, if none are specified than
 #' the original dataset passed to `e_charts` is used.
+#' @param value The column to map to the parameter.
+#' 
+#' @section Functions:
+#' 
+#' - `e_add_nested`: Adds nested data, e.g.: 
+#' `e_add_nested("itemStyle", color, fontBold)` creates
+#' `{itemStyle: {color: 'red', fontBold: 'bold'}}`.
+#' - `e_add_unnested`: Adds unnested data, e.g.:
+#' `e_add_unnested("symbolSize", size)` creates
+#' `{symbolSize: 4}`.
 #'
 #' @details For instance, \code{\link{e_funnel}} lets you pass \code{values} and \code{labels}
 #' (from your initial data.frame) which corresponds to \code{name} and \code{value} in the
@@ -28,7 +38,7 @@
 #' funnel |>
 #'   e_charts() |>
 #'   e_funnel(value, stage) |>
-#'   e_add("itemStyle", color)
+#'   e_add_nested("itemStyle", color)
 #'
 #' # Heatmap can take nested label
 #' # https://echarts.apache.org/en/option.html#series-heatmap.data
@@ -51,13 +61,22 @@
 #'   e_charts(x) |>
 #'   e_heatmap(y, z) |>
 #'   e_visual_map(z) |>
-#'   e_add(
+#'   e_add_nested(
 #'     "label",
 #'     show,
 #'     fontStyle
 #'   )
+#' 
+#' @name nesting
 #' @export
 e_add <- function(e, param, ..., .serie = NULL, .data = NULL) {
+  .Deprecated("e_add_nested", "echarts4r")
+  e_add_nested(e, param, ..., .serie = NULL, .data = NULL)
+}
+
+#' @rdname nesting
+#' @export 
+e_add_nested <- function(e, param, ..., .serie = NULL, .data = NULL) {
   if (missing(e) || missing(param)) {
     stop("missing e or param", call. = FALSE)
   }
@@ -82,6 +101,39 @@ e_add <- function(e, param, ..., .serie = NULL, .data = NULL) {
       } else {
         for(k in seq_along(e$x$opts$options[[i]]$series)){
           e$x$opts$options[[i]]$series[[k]]$data[[j]][[param]] <- data[[j]]
+        }
+      }
+    }
+  }
+  e
+}
+
+#' @rdname nesting
+#' @export 
+e_add_unnested <- function(e, param, value, .serie = NULL, .data = NULL) {
+  if (missing(e) || missing(param) || missing(value)) {
+    stop("missing e, param, or value", call. = FALSE)
+  }
+
+  if(is.null(.data))
+    ds <- e$x$data
+  else
+    ds <- map_grps_(.data, e$x$tl)
+
+  for (i in seq_along(ds)) {
+    data <- ds[[i]] |>
+      dplyr::pull({{ value }}) |> 
+      unname()
+
+    for (j in seq_along(data)) {
+      if(!is.null(.serie) && .serie != j)
+        next
+      
+      if (!e$x$tl) {
+        e$x$opts$series[[i]]$data[[j]][[param]] <- data[j]
+      } else {
+        for(k in seq_along(e$x$opts$options[[i]]$series)){
+          e$x$opts$options[[i]]$series[[k]]$data[[j]][[param]] <- data[j]
         }
       }
     }